Removed self-absorbing Markov chains impl.
Previous commit allowed the MDP to be a self-absorbing markov chain.
This led to some odd upper limits on Q-Values of terminal states equal
to (reward / (gamma - 1) = -( Q(s,a) ). In my case for an immediate
reward of 1, this led to a limit of 10 for gamma = 0.1.
This fix means that terminal states have no Q values for actions at
all other than those initialised.
GridWorld will immediately terminate upon current state reaching a
terminal state.
This can be fixed in representation by accounting for terminal states
and refusing to display Max at Q(st,at).